clean up masked array module
